HotelCenter/requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==36.0.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.3
django-birthday==0.1.4
django-cors-headers==3.11.0
django-filter==21.1
django-nose==1.4.7
django-render-block==0.9.1
django-templated-email==3.0.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flower==1.1.0
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
humanize==4.2.3
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.1.0
msgpack==1.0.3
mysql-client==0.0.1
mysqlclient==2.1.0
ndg-httpsclient==0.5.1
nose==1.3.7
numpy==1.22.3
oauthlib==3.2.0
outcome==1.1.0
packaging==21.3
Pillow==9.0.1
postgres==4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pusher==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
redis==4.3.3
requests==2.27.1
requests-oauthlib==1.3.1
selenium==4.1.5
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tornado==6.2
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
wsproto==1.1.0
zope.interface==5.4.0
